Refactor: Move font handling to lib package (shield-generation)
Moves font-related constants, types, and functions from
`utils/badges.go` to a new `lib/fonts.go` file. This centralizes
font management and improves code organization.

Additionally, updates the font existence test to use the new
`lib.FontMap` and corrects the path to the project directory.
Finally, updates the import paths in files using font utilities.
