+++ utils/badges.go
@@ -3,80 +3,15 @@ package utils
 import (
 	"fmt"
 	"log"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
+
+	"github.com/jasonlovesdoggo/abacus/lib"
 
 	"github.com/jasonlovesdoggo/abacus/lib/badge"
 
 	"github.com/gin-gonic/gin"
 )
-
-// FontInfo contains details about a font
-type FontInfo struct {
-	FileName   string
-	FontFamily string
-}
-
-// Map of supported fonts with their file names and CSS font-family values
-var fontMap = map[string]FontInfo{
-	"verdana": {
-		FileName:   "Verdana.ttf",
-		FontFamily: "Verdana,DejaVu Sans,sans-serif",
-	},
-	"verdana-bold": {
-		FileName:   "Verdana_Bold.ttf",
-		FontFamily: "Verdana Bold,DejaVu Sans,sans-serif",
-	},
-	"verdana-bold-italic": {
-		FileName:   "Verdana_Bold_Italic.ttf",
-		FontFamily: "Verdana Bold Italic,DejaVu Sans,sans-serif",
-	},
-	"arial": {
-		FileName:   "Arial.ttf",
-		FontFamily: "Arial,Helvetica,sans-serif",
-	},
-	"arial-bold": {
-		FileName:   "Arial_Bold.ttf",
-		FontFamily: "Arial Bold,Helvetica,sans-serif",
-	},
-	"arial-italic": {
-		FileName:   "Arial_Italic.ttf",
-		FontFamily: "Arial Italic,Helvetica,sans-serif",
-	},
-	"arial-bold-italic": {
-		FileName:   "Arial_Bold_Italic.ttf",
-		FontFamily: "Arial Bold Italic,Helvetica,sans-serif",
-	},
-	"courier-new": {
-		FileName:   "Courier_New.ttf",
-		FontFamily: "Courier New,Courier,monospace",
-	},
-	"jetbrains-mono": {
-		FileName:   "JetbrainsMono.ttf",
-		FontFamily: "JetBrains Mono,Courier New,monospace",
-	},
-}
-
-func getFontFilePath(font string) (string, string, error) {
-	execDir, _ := os.Getwd()
-
-	fontInfo, exists := fontMap[font]
-	if !exists {
-		// Default to Verdana if font name not found
-		fontInfo = fontMap["verdana"]
-	}
-
-	fontPath := filepath.Join(execDir, "assets", "fonts", fontInfo.FileName)
-
-	// Check if font file exists
-	if _, err := os.Stat(fontPath); os.IsNotExist(err) {
-		return "", "", fmt.Errorf("font file not found: %s", fontPath)
-	}
-
-	return fontPath, fontInfo.FontFamily, nil
-}
 
 func GenerateBadge(c *gin.Context, count int64) ([]byte, error) {
 	bgColor := c.DefaultQuery("bgcolor", "007ec6")
@@ -105,7 +40,7 @@ func GenerateBadge(c *gin.Context, count int64) ([]byte, error) {
 	}
 
 	// Get font path and font family
-	filePath, fontFamily, err := getFontFilePath(font)
+	filePath, fontFamily, err := lib.GetFontFilePath(font)
 	if err != nil {
 		log.Printf("Error: Failed to get font file path: %v", err)
 		return nil, fmt.Errorf("font error: %w", err)
